+++ pi_audio.h
@@ -1,5 +1,6 @@
 #ifndef _PI_AUDIO_H
 #define _PI_AUDIO_H
+#include <stdint.h>
 
 #define MAX_CHANNELS 2
 
@@ -8,7 +9,7 @@
 #define BACKLOG         5
 
 /* How many frames of audio the master has played. */
-extern unsigned long master_frames_played;
+extern uint32_t master_frames_played;
 
 /* Strings representing port values; max is "65536", so 10 bytes is enough. */
 char stream_port_str [10];
